network namespace moved to the main
