+++ dev/Interfaces/CMakeLists.txt
@@ -26,7 +26,7 @@ set(INTERFACE_HEADER_FILES
                     Interfaces/controller_interface.h
                     Interfaces/model_interface.h
                     Interfaces/view_interface.h 
-                    Interfaces/main_interface.h)
+                    Interfaces/main_interface.h "Interfaces/network_interface.h")
 set(INTERFACE_SOURCE_FILES 
                     Interfaces/src/controller_interface.cpp
                     Interfaces/src/model_interface.cpp
