+++ dev/Network/CMakeLists.txt
@@ -1,9 +1,11 @@
 set(NETWORK_INTERFACE_HEADERS
+                    network_main.h
                     messages.h
                     system_interfaces/network_interface.h
                     system_interfaces/types/ipv4_addr.h
                     system_interfaces/types/network_status_state.h)
 set(NETWORK_INTERFACE_SOURCE
+                    network_main.cpp
                     src/connections.cpp
                     src/network_interface.cpp
                     src/windows_network_interface.cpp
